requirements.txt
opencv-python
Pillow
scikit-learn
matplotlib
pandas
numpy